requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
joblib==1.3.2
lazy_loader==0.3
msgpack==1.0.7
mysqlclient==2.2.1
packaging==23.2
Pillow==10.1.0
pip==24.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.24.4
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rptree==0.1.1
scikit-learn==1.3.2
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0